Move SYCL devicelib to toplevel.
