+++ libdevice/fallback-cmath-fp64.cpp
@@ -7,205 +7,205 @@
 //===----------------------------------------------------------------------===//
 #ifdef __SYCL_DEVICE_ONLY__
 #include "device_math.h"
-extern "C" {
-SYCL_EXTERNAL
-float __devicelib_scalbnf(float x, int n) {
-  return __spirv_ocl_ldexp(x, n);
-}
 
+#ifdef __cplusplus
+extern "C" {
+#endif  // __cplusplus
 SYCL_EXTERNAL
-float __devicelib_logf(float x) {
+double __devicelib_log(double x) {
   return __spirv_ocl_log(x);
 }
 
 SYCL_EXTERNAL
-float __devicelib_expf(float x) {
+double __devicelib_exp(double x) {
   return __spirv_ocl_exp(x);
 }
 
 SYCL_EXTERNAL
-float __devicelib_frexpf(float x, int *exp) {
+double __devicelib_frexp(double x, int *exp) {
   return __spirv_ocl_frexp(x, exp);
 }
 
 SYCL_EXTERNAL
-float __devicelib_ldexpf(float x, int exp) {
+double __devicelib_ldexp(double x, int exp) {
   return __spirv_ocl_ldexp(x, exp);
 }
 
 SYCL_EXTERNAL
-float __devicelib_log10f(float x) {
+double __devicelib_log10(double x) {
   return __spirv_ocl_log10(x);
 }
 
 SYCL_EXTERNAL
-float __devicelib_modff(float x, float *intpart) {
+double __devicelib_modf(double x, double *intpart) {
   return __spirv_ocl_modf(x, intpart);
 }
 
 SYCL_EXTERNAL
-float __devicelib_exp2f(float x) {
+double __devicelib_exp2(double x) {
   return __spirv_ocl_exp2(x);
 }
 
 SYCL_EXTERNAL
-float __devicelib_expm1f(float x) {
+double __devicelib_expm1(double x) {
   return __spirv_ocl_expm1(x);
 }
 
 SYCL_EXTERNAL
-int __devicelib_ilogbf(float x) {
+int __devicelib_ilogb(double x) {
   return __spirv_ocl_ilogb(x);
 }
 
 SYCL_EXTERNAL
-float __devicelib_log1pf(float x) {
+double __devicelib_log1p(double x) {
   return __spirv_ocl_log1p(x);
 }
 
 SYCL_EXTERNAL
-float __devicelib_log2f(float x) {
+double __devicelib_log2(double x) {
   return __spirv_ocl_log2(x);
 }
 
 SYCL_EXTERNAL
-float __devicelib_logbf(float x) {
+double __devicelib_logb(double x) {
   return __spirv_ocl_logb(x);
 }
 
 SYCL_EXTERNAL
-float __devicelib_sqrtf(float x) {
+double __devicelib_sqrt(double x) {
   return __spirv_ocl_sqrt(x);
 }
 
 SYCL_EXTERNAL
-float __devicelib_cbrtf(float x) {
+double __devicelib_cbrt(double x) {
   return __spirv_ocl_cbrt(x);
 }
 
 SYCL_EXTERNAL
-float __devicelib_hypotf(float x, float y) {
+double __devicelib_hypot(double x, double y) {
   return __spirv_ocl_hypot(x, y);
 }
 
 SYCL_EXTERNAL
-float __devicelib_erff(float x) {
+double __devicelib_erf(double x) {
   return __spirv_ocl_erf(x);
 }
 
 SYCL_EXTERNAL
-float __devicelib_erfcf(float x) {
+double __devicelib_erfc(double x) {
   return __spirv_ocl_erfc(x);
 }
 
 SYCL_EXTERNAL
-float __devicelib_tgammaf(float x) {
+double __devicelib_tgamma(double x) {
   return __spirv_ocl_tgamma(x);
 }
 
 SYCL_EXTERNAL
-float __devicelib_lgammaf(float x) {
+double __devicelib_lgamma(double x) {
   return __spirv_ocl_lgamma(x);
 }
 
 SYCL_EXTERNAL
-float __devicelib_fmodf(float x, float y) {
+double __devicelib_fmod(double x, double y) {
   return __spirv_ocl_fmod(x, y);
 }
 
 SYCL_EXTERNAL
-float __devicelib_remainderf(float x, float y) {
+double __devicelib_remainder(double x, double y) {
   return __spirv_ocl_remainder(x, y);
 }
 
 SYCL_EXTERNAL
-float __devicelib_remquof(float x, float y, int *q) {
+double __devicelib_remquo(double x, double y, int *q) {
   return __spirv_ocl_remquo(x, y, q);
 }
 
 SYCL_EXTERNAL
-float __devicelib_nextafterf(float x, float y) {
+double __devicelib_nextafter(double x, double y) {
   return __spirv_ocl_nextafter(x, y);
 }
 
 SYCL_EXTERNAL
-float __devicelib_fdimf(float x, float y) {
+double __devicelib_fdim(double x, double y) {
   return __spirv_ocl_fdim(x, y);
 }
 
 SYCL_EXTERNAL
-float __devicelib_fmaf(float x, float y, float z) {
+double __devicelib_fma(double x, double y, double z) {
   return __spirv_ocl_fma(x, y, z);
 }
 
 SYCL_EXTERNAL
-float __devicelib_sinf(float x) {
+double __devicelib_sin(double x) {
   return __spirv_ocl_sin(x);
 }
 
 SYCL_EXTERNAL
-float __devicelib_cosf(float x) {
+double __devicelib_cos(double x) {
   return __spirv_ocl_cos(x);
 }
 
 SYCL_EXTERNAL
-float __devicelib_tanf(float x) {
+double __devicelib_tan(double x) {
   return __spirv_ocl_tan(x);
 }
 
 SYCL_EXTERNAL
-float __devicelib_powf(float x, float y) {
+double __devicelib_pow(double x, double y) {
   return __spirv_ocl_pow(x, y);
 }
 
 SYCL_EXTERNAL
-float __devicelib_acosf(float x) {
+double __devicelib_acos(double x) {
   return __spirv_ocl_acos(x);
 }
 
 SYCL_EXTERNAL
-float __devicelib_asinf(float x) {
+double __devicelib_asin(double x) {
   return __spirv_ocl_asin(x);
 }
 
 SYCL_EXTERNAL
-float __devicelib_atanf(float x) {
+double __devicelib_atan(double x) {
   return __spirv_ocl_atan(x);
 }
 
 SYCL_EXTERNAL
-float __devicelib_atan2f(float x, float y) {
+double __devicelib_atan2(double x, double y) {
   return __spirv_ocl_atan2(x, y);
 }
 
 SYCL_EXTERNAL
-float __devicelib_coshf(float x) {
-  return  __spirv_ocl_cosh(x);
+double __devicelib_cosh(double x) {
+  return __spirv_ocl_cosh(x);
 }
 
 SYCL_EXTERNAL
-float __devicelib_sinhf(float x) {
+double __devicelib_sinh(double x) {
   return __spirv_ocl_sinh(x);
 }
 
 SYCL_EXTERNAL
-float __devicelib_tanhf(float x) {
+double __devicelib_tanh(double x) {
   return __spirv_ocl_tanh(x);
 }
 
 SYCL_EXTERNAL
-float __devicelib_acoshf(float x) {
+double __devicelib_acosh(double x) {
   return __spirv_ocl_acosh(x);
 }
 
 SYCL_EXTERNAL
-float __devicelib_asinhf(float x) {
+double __devicelib_asinh(double x) {
   return __spirv_ocl_asinh(x);
 }
 
 SYCL_EXTERNAL
-float __devicelib_atanhf(float x) {
+double __devicelib_atanh(double x) {
   return __spirv_ocl_atanh(x);
 }
-}
-#endif
+#ifdef __cplusplus
+} // extern "C"
+#endif  // __cplusplus
+#endif  // __SYCL_DEVICE_ONLY__
