+++ sycl/source/CMakeLists.txt
@@ -185,4 +185,3 @@ install(TARGETS ${SYCL_RT_LIBS}
     ARCHIVE DESTINATION "lib${LLVM_LIBDIR_SUFFIX}" COMPONENT sycl
     LIBRARY DESTINATION "lib${LLVM_LIBDIR_SUFFIX}" COMPONENT sycl
     RUNTIME DESTINATION "bin" COMPONENT sycl)
-add_subdirectory(detail/devicelib)
